Improve error reporting for network pods
If the pod isn't scheduled, capture the reason and the pod status rather
than vaguely reporting that it timed out.

Also we've seen tcpdump fail to run due to permissions but the output
doesn't indicate the root cause. Report the content of the pod log if
the captured container status message is empty.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>